8/3일 푸쉬

+++ src/main/java/com/kh/board/controller/BoardRepDelServlet.java
@@ -7,8 +7,11 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+
 import com.kh.board.model.service.BoardService;
+import com.kh.board.model.vo.Reply;
 import com.kh.common.util.MyHttpServlet;
+import com.kh.member.model.vo.User;
 
 @WebServlet("/board/repdel")
 public class BoardRepDelServlet extends MyHttpServlet {
@@ -27,11 +30,19 @@ public class BoardRepDelServlet extends MyHttpServlet {
 		try {
 			int boardNo = Integer.parseInt(req.getParameter("boardNo"));
 			int replyNo = Integer.parseInt(req.getParameter("replyNo"));
+			
+			Reply r =bs.searchReply(replyNo);
+			User loginUser = getSessionUser(req);
+			if (loginUser!=null && !loginUser.getUser_id().equals(r.getBoard_reply_writer())) {
+				sendCommonPage("세션이 만료되었습니다.", "/board/view?boardNo="+boardNo, req, resp);
+				return;
+			}
+			
 			int result = bs.deleteReply(replyNo);
 			if (result > 0) {
-				sendCommonPage("success", "/board/view?boardNo=" + boardNo, req, resp);
+				sendCommonPage("댓글 삭제에 성공하셨습니다.", "/board/view?boardNo=" + boardNo, req, resp);
 			} else {
-				sendCommonPage("fail", "/board/view?boardNo=" + boardNo, req, resp);
+				sendCommonPage("댓글 삭제에 실패하셨습니다.", "/board/view?boardNo=" + boardNo, req, resp);
 			}
 		} catch (Exception e) {
 			e.printStackTrace();
